requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.2
Django==3.2.13
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
mypy==0.942
mypy-extensions==0.4.3
pycryptodomex==3.14.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
telepot==12.6
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
